Refactor to remove legacy module
