+++ core/__init__.py
@@ -1,0 +1,4 @@
+# Refactored on 2024-06-06 to remove legacy coupling
+from .market_data import Bar, OHLCCollector, data_stream
+from .indicators_vectorized import compute_rsi, atr, compute_adx
+
