+++ strategy/__init__.py
@@ -1,0 +1,5 @@
+# Refactored on 2024-06-06 to remove legacy coupling
+from .entry import BounceEntry, Signal, is_reversal_candle
+from .dca import SmartDCA, DCAFilters
+from .manager import PositionManager
+
