+++ strategy/entry.py
@@ -1,3 +1,4 @@
+# Refactored on 2024-06-06 to remove legacy coupling
 from __future__ import annotations
 
 from enum import Enum
@@ -5,16 +6,13 @@ from typing import Sequence, Tuple
 import statistics
 
 
-
-
-class EntrySignal(Enum):
-    NO = 0
+class Signal(Enum):
     LONG = 1
     SHORT = -1
+    FLAT = 0
 
 
 def is_reversal_candle(open_p: float, high: float, low: float, close: float) -> bool:
-    """Return ``True`` if candle looks like a reversal (pin-bar or belt hold)."""
     rng = high - low
     if rng <= 0:
         return False
@@ -38,56 +36,47 @@ class BounceEntry:
         bb_params: Tuple[float | None, float | None],
         rsi_params: Tuple[float, float, float],
         adx: float,
-    ) -> EntrySignal:
-        """Return entry signal based on indicator values."""
+    ) -> Signal:
         lower, upper = bb_params
         rsi_val, rsi_low, rsi_high = rsi_params
         if lower is None or upper is None:
-            return EntrySignal.NO
-        direction = EntrySignal.NO
+            return Signal.FLAT
+        direction = Signal.FLAT
         if bar.close <= lower:
-            direction = EntrySignal.LONG
+            direction = Signal.LONG
         elif bar.close >= upper:
-            direction = EntrySignal.SHORT
+            direction = Signal.SHORT
         else:
-            return EntrySignal.NO
+            return Signal.FLAT
 
-        if direction == EntrySignal.LONG and rsi_val >= rsi_low:
-            return EntrySignal.NO
-        if direction == EntrySignal.SHORT and rsi_val <= rsi_high:
-            return EntrySignal.NO
+        if direction == Signal.LONG and rsi_val >= rsi_low:
+            return Signal.FLAT
+        if direction == Signal.SHORT and rsi_val <= rsi_high:
+            return Signal.FLAT
 
         if adx >= 25:
-            return EntrySignal.NO
+            return Signal.FLAT
 
         if not is_reversal_candle(bar.open, bar.high, bar.low, bar.close):
-            return EntrySignal.NO
+            return Signal.FLAT
 
         if len(volume_window) < 2:
-            return EntrySignal.NO
+            return Signal.FLAT
         avg_vol = statistics.mean(volume_window[:-1]) if len(volume_window) > 1 else 0.0
         if avg_vol <= 0 or bar.volume < 2 * avg_vol:
-            return EntrySignal.NO
+            return Signal.FLAT
 
         return direction
 
     @staticmethod
-    def check(
-        bar,
-        volume_window: Sequence[float],
-        close_window: Sequence[float],
-        params: object | dict,
-    ) -> EntrySignal | None:
-        """Return ``EntrySignal`` if all conditions are met."""
+    def check(bar, volume_window: Sequence[float], close_window: Sequence[float], params: object | dict) -> Signal | None:
         dev = getattr(params, "bb_dev", None)
         if isinstance(params, dict):
             dev = params.get("bb_dev", dev)
         bb_dev = dev if dev is not None else 2.0
-
         if len(close_window) < 20:
             return None
-
-        from legacy.core import indicators
+        from app import indicators
 
         lower, _, upper = indicators.bollinger(list(close_window), 20, bb_dev)
         rsi_v = indicators.rsi(list(close_window), 14)
@@ -98,4 +87,4 @@ class BounceEntry:
             (rsi_v, 30.0, 70.0),
             0.0,
         )
-        return sig if sig is not EntrySignal.NO else None
+        return sig if sig is not Signal.FLAT else None
